requirements.txt
streamlit==1.28.1
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.25.2